# support R.__ placeholder in curry functions