--- a/src/curryN.js
+++ b/src/curryN.js
@@ -1,25 +1,39 @@
-var _concat = require('./internal/_concat');
+var __ = require('../__');
 var _curry2 = require('./internal/_curry2');
 var _noArgsException = require('./internal/_noArgsException');
+var _slice = require('./internal/_slice');
 var arity = require('./arity');
 
 
 /**
- * Creates a new version of `fn` with given arity that, when invoked,
- * will return either:
- * - A new function ready to accept one or more of `fn`'s remaining arguments, if all of
- * `fn`'s expected arguments have not yet been provided
- * - `fn`'s result if all of its expected arguments have been provided
+ * Returns a curried equivalent of the provided function, with the
+ * specified arity. The curried function has two unusual capabilities.
+ * First, its arguments needn't be provided one at a time. If `g` is
+ * `R.curryN(3, f)`, the following are equivalent:
  *
- * This function is useful in place of `curry`, when the arity of the
- * function to curry cannot be determined from its signature, e.g. if it's
- * a variadic function.
+ *   - `g(1)(2)(3)`
+ *   - `g(1)(2, 3)`
+ *   - `g(1, 2)(3)`
+ *   - `g(1, 2, 3)`
+ *
+ * Secondly, the special placeholder value `R.__` may be used to specify
+ * "gaps", allowing partial application of any combination of arguments,
+ * regardless of their positions. If `g` is as above and `_` is `R.__`,
+ * the following are equivalent:
+ *
+ *   - `g(1, 2, 3)`
+ *   - `g(_, 2, 3)(1)`
+ *   - `g(_, _, 3)(1)(2)`
+ *   - `g(_, _, 3)(1, 2)`
+ *   - `g(_, 2)(1)(3)`
+ *   - `g(_, 2)(1, 3)`
+ *   - `g(_, 2)(_, 3)(1)`
  *
  * @func
  * @memberOf R
  * @category Function
  * @sig Number -> (* -> a) -> (* -> a)
- * @param {Number} fnArity The arity for the returned function.
+ * @param {Number} length The arity for the returned function.
  * @param {Function} fn The function to curry.
  * @return {Function} A new, curried function.
  * @see R.curry
@@ -32,18 +46,35 @@ var arity = require('./arity');
  *      var curriedAddFourNumbers = R.curryN(4, addFourNumbers);
  *      var f = curriedAddFourNumbers(1, 2);
  *      var g = f(3);
- *      g(4);//=> 10
+ *      g(4); //=> 10
  */
 module.exports = _curry2(function curryN(length, fn) {
-    return (function recurry(args) {
-        return arity(Math.max(length - (args && args.length || 0), 0), function() {
-            if (arguments.length === 0) { throw _noArgsException(); }
-            var newArgs = _concat(args, arguments);
-            if (newArgs.length >= length) {
-                return fn.apply(this, newArgs);
-            } else {
-                return recurry(newArgs);
+    return arity(length, function() {
+        var n = arguments.length;
+        if (n === 0) {
+            throw _noArgsException();
+        }
+        var shortfall = length - n;
+        var idx = n;
+        while (idx--) {
+            if (arguments[idx] === __) {
+                shortfall += 1;
             }
-        });
-    }([]));
+        }
+        if (shortfall <= 0) {
+            return fn.apply(this, arguments);
+        } else {
+            var initialArgs = _slice(arguments);
+            return curryN(shortfall, function() {
+                var currentArgs = _slice(arguments);
+                var combinedArgs = [];
+                var idx = -1;
+                while (++idx < n) {
+                    var val = initialArgs[idx];
+                    combinedArgs[idx] = val === __ ? currentArgs.shift() : val;
+                }
+                return fn.apply(this, combinedArgs.concat(currentArgs));
+            });
+        }
+    });
 });