--- a/src/lift.js
+++ b/src/lift.js
@@ -15,12 +15,12 @@ var liftN = require('./liftN');
  * @return {Function} The function `fn` applicable to mappable objects.
  * @example
  *
- *     var madd3 = R.lift(R.curryN(3, function(a, b, c) {
+ *     var madd3 = R.lift(R.curry(function(a, b, c) {
  *         return a + b + c;
  *     }));
  *     madd3([1,2,3], [1,2,3], [1]); //=> [3, 4, 5, 4, 5, 6, 5, 6, 7]
  *
- *     var madd5 = R.lift(R.curryN(5, function(a, b, c, d, e) {
+ *     var madd5 = R.lift(R.curry(function(a, b, c, d, e) {
  *         return a + b + c + d + e;
  *     }));
  *     madd5([1,2], [3], [4, 5], [6], [7, 8]); //=> [21, 22, 22, 23, 22, 23, 23, 24]