--- a/src/lte.js
+++ b/src/lte.js
@@ -1,4 +1,4 @@
-var op = require('./op');
+var _curry2 = require('./internal/_curry2');
 
 
 /**
@@ -11,8 +11,6 @@ var op = require('./op');
  * @param {Number} a
  * @param {Number} b
  * @return {Boolean} a <= b
- * @note Operator: Since this is a non-commutative infix operator converted to prefix, it can
- *                 be curried right by explicitly passing `R.__` for its first argument.
  * @example
  *
  *      R.lte(2, 6); //=> true
@@ -20,6 +18,5 @@ var op = require('./op');
  *      R.lte(2, 2); //=> true
  *      R.lte(R.__, 2)(1); //=> true
  *      R.lte(2)(10); //=> true
- *      R.lte(R.__)(5, 4) // => true
  */
-module.exports = op(function lte(a, b) { return a <= b; });
+module.exports = _curry2(function lte(a, b) { return a <= b; });