--- a/test/contains.js
+++ b/test/contains.js
@@ -23,12 +23,11 @@ describe('contains', function() {
     });
 
     it('is curried like a binary operator, that accepts an inital placeholder', function() {
-        var fbb = ['foo', 'bar', 'baz'];
-        assert(typeof R.contains(R.__) === 'function');
-        assert(typeof R.contains(R.__)('bar') === 'function');
-        assert(R.contains(R.__)(fbb)('bar'));
-        assert(R.contains(R.__, fbb)('bar'));
-        assert(R.contains(R.__)(fbb, 'bar'));
+        var isDigit = R.contains(R.__, ['0', '1', '2', '3', '4', '5', '6', '7', '8', '9']);
+        assert.strictEqual(typeof isDigit, 'function');
+        assert.strictEqual(isDigit('0'), true);
+        assert.strictEqual(isDigit('1'), true);
+        assert.strictEqual(isDigit('x'), false);
     });
 
     it('throws on zero arguments', function() {