… dist/ramda.js
… src/op.js
… test/internal/_curry3.js
… test/op.js