See #817

Changes:

  - add internal `_curryN` function, similar to `op` but capable of currying functions of arbitrary arity;
  - define `curryN` as `_curryN(2, _curryN)`;
  - update `_checkForMethod` to return a function with the appropriate arity;
  - replace occurrences of `_curry2`, `_curry3`, and `op` with `curry`; and
  - remove `_curry2`, `_curry3`, and `op`.

Reasons to like these changes:

  - it will put an end to ad hoc pull requests adding "op-style" currying to one or two functions at a time;
  - it improves consistency✝;
  - partial application is not *only* useful for binary functions;
  - it removes the distinction between "operator" functions and other functions, simplifying Ramda's mental model and making `R.__` easier to explain; and
  - it further promotes partial application, Ramda's key feature.

Reasons to dislike these changes:

  - it's a breaking change✝✝; and
  - it increases the amount of work required to apply a curried function.


✝ Currently commutative functions do not support `R.__`, as it's not useful. For consistency, though, they should. One could then say that `R.curry(f)(R.__, b)(a)` is equivalent to `f(a, b)` for any binary function `f`, in the same way one can currently say that `R.curry(f)(a)(b)` is equivalent to `f(a, b)` for any binary function `f`. (All the discussion in Fantasy Land seems to have got me thinking in terms of laws and whatnot.)


✝✝ Currently `R.divide(R.__)` is equivalent to `R.flip(R.divide)`, which has always seemed to me like too much magic. I doubt anyone is using this form, as the `R.flip` version is much clearer.