@CrossEye 2015-02-17
I talked about this a while ago as a good idea.  I saw #817, and agreed with it.

My only hesitation is whether this will do any significant damage to performance.  I know that's not your focus.  Nor is it mine.  And it's not my strong suit.  But the nastiness of the current Balkanized curries comes out of a need to improve performance.  I don't think you were participating then, but soon after we announced Ramda, the author of a popular JS library posted some performance tests of some of our key functions and tweeted that pre-currying is a huge performance killer.  Since we hadn't even looked at performance at that point, we took it to heart and went through two rounds of radical surgery to get versions that had APIs we could live with and performance that didn't disappoint.  (That author has now released a version of his library with an API that looks an awful lot like ours, including the pre-currying, so he must have had some change of heart.) 

So, have you compared benchmarks before and after?  For the improved consistency and flexibility, I would not mind if we lost a little performance, but I would want to be sure it's not crippling.