@davidchambers 2015-02-17
[jsPerf][1] is currently offline. Do you have a test case in mind, @CrossEye?

To provide a (very) rough idea of the performance, running the test suite of a project which uses Ramda extensively took 2s before this change and 3s afterwards.


[1]: https://jsperf.com/