@buzzdecafe 2015-02-17
@davidchambers you should be able to run `grunt benchmark` in your branch and compare the output in bench/report to running it on master.