@CrossEye 2015-02-17
I was thinking about the difference between the outputs of `grunt benchmark` run in HEAD versus in this branch.