@kedashoe 2015-02-17
> remove _curry2, _curry3

Why not create placeholder versions of these? I don't personally care much about speed but it is simple enough to write https://gist.github.com/kedashoe/e0184a5b91e63084299b. Since all internal ramda functions used to use these, maybe it is worthwhile?