@buzzdecafe 2015-02-17
@kedashoe let's see where we stand first. If &delta; &lt; &epsilon; then there's no need for such optimizations. If &delta; > &epsilon; then it may be worth optimizing as suggested in that gist.