@kedashoe 2015-02-17
> let's see where we stand first. If δ < ε then there's no need for such optimizations. If δ > ε then it may be worth optimizing as suggested in that gist.

Fair enough, although there is no trade-off, right? Placeholder currying is a trade-off in that you get more features at the cost of reduced speed. This is just better speed with no cost?

Also it looks like this curry passes along additional arguments, which the old curry2 and curry3 did not do, changing a bit all internal functions that used curry2 and curry3 (the gist also does not pass along arguments, but can be made to do so if that is the direction this is headed).