@davidchambers 2015-02-18
I've updated this pull request to use @kedashoe's optimizations. Here are the updated numbers:

| test             | master ops/sec | placeholder ops/sec |
| ---------------- | --------------:| -------------------:|
| `concat(s1, s2)` |     10,247,626 |           9,428,834 |
| `concat(s1)(s2)` |      1,710,294 |           7,313,137 |
| `concatS1(s2)`   |      5,240,847 |          10,860,689 |

Here's the full output:

```
20:34 ~/github.com/ramda/ramda (placeholder)
$ grunt benchmark
Running "benchmark:all" (benchmark) task

Running suite concat [lib/bench/concat.bench.js]...
>> _(s1).concat(s2) x 1,448,262 ops/sec ±2.42% (99 runs sampled)
>> concat(s1, s2) x 9,428,834 ops/sec ±3.12% (89 runs sampled)
>> concat(s1)(s2) x 7,313,137 ops/sec ±5.73% (81 runs sampled)
>> concatS1(s2) x 10,860,689 ops/sec ±3.11% (96 runs sampled)
Fastest test is concatS1(s2) at 1.15x faster than concat(s1, s2)

Running suite curry [lib/bench/curry.bench.js]...
>> _x4(100) x 11,949,741 ops/sec ±3.10% (91 runs sampled)
>> mult4(100) x 1,539,249 ops/sec ±2.98% (91 runs sampled)
>> manual(100) x 54,511,421 ops/sec ±2.72% (86 runs sampled)
Fastest test is manual(100) at 4.6x faster than _x4(100)

Running suite forEach [lib/bench/each.bench.js]...
>> _.each(nums, x2) x 393,624 ops/sec ±3.24% (89 runs sampled)
>> forEach(x2, nums) x 434,199 ops/sec ±0.76% (98 runs sampled)
>> forEach(x2)(nums) x 369,324 ops/sec ±5.60% (85 runs sampled)
Fastest test is forEach(x2, nums) at 1.18x faster than forEach(x2)(nums) and _.each(nums, x2)

Running suite filter where [lib/bench/filter-where.bench.js]...
>> _.filter(objs, {x: []}) x 105,438 ops/sec ±4.80% (90 runs sampled)
>> filter(where({x: isEmpty}), objs) x 272,448 ops/sec ±8.17% (72 runs sampled)
>> filter(where({x: isEmpty}))(objs) x 317,189 ops/sec ±6.36% (83 runs sampled)
>> filterEmptyX(objs) x 635,131 ops/sec ±2.76% (92 runs sampled)
>> _.filter(objs, {y: false}) x 945,481 ops/sec ±1.52% (93 runs sampled)
>> filter(where({y: false}), objs) x 460,322 ops/sec ±2.19% (89 runs sampled)
>> filter(where({y: false}))(objs) x 455,764 ops/sec ±3.97% (91 runs sampled)
>> filterFalseY(objs) x 1,632,214 ops/sec ±2.06% (90 runs sampled)
Fastest test is filterFalseY(objs) at 1.73x faster than _.filter(objs, {y: false})

Running suite filter [lib/bench/filter.bench.js]...
>> _.filter(nums, isEven) x 2,496,910 ops/sec ±1.89% (92 runs sampled)
>> filter(isEven, nums) x 2,415,417 ops/sec ±1.84% (92 runs sampled)
>> filter(isEven)(nums) x 2,137,426 ops/sec ±3.43% (93 runs sampled)
>> filterEven(nums) x 2,587,186 ops/sec ±2.06% (94 runs sampled)
Fastest test is filterEven(nums) at 1.04x faster than _.filter(nums, isEven)

Running suite find where [lib/bench/find-where.bench.js]...
>> _.find(objs, {x: []}) x 605,994 ops/sec ±4.49% (90 runs sampled)
>> find(where({x: isEmpty}), objs) x 518,733 ops/sec ±0.66% (93 runs sampled)
>> find(where({x: isEmpty}))(objs) x 513,727 ops/sec ±3.60% (88 runs sampled)
>> findEmptyX(objs) x 2,293,058 ops/sec ±1.16% (97 runs sampled)
>> _.find(objs, {y: false}) x 1,263,956 ops/sec ±3.98% (92 runs sampled)
>> find(where({y: false}), objs) x 641,623 ops/sec ±2.88% (94 runs sampled)
>> find(where({y: false}))(objs) x 604,415 ops/sec ±3.41% (92 runs sampled)
>> findFalseY(objs) x 6,681,018 ops/sec ±1.94% (97 runs sampled)
Fastest test is findFalseY(objs) at 2.9x faster than findEmptyX(objs)

Running suite find [lib/bench/find.bench.js]...
>> _.find(nums, isZero) x 6,521,106 ops/sec ±0.63% (98 runs sampled)
>> find(isZero, nums) x 11,210,923 ops/sec ±2.95% (94 runs sampled)
>> find(isZero)(nums) x 8,474,279 ops/sec ±2.35% (92 runs sampled)
>> findZero(nums) x 11,714,365 ops/sec ±3.01% (90 runs sampled)
Fastest test is findZero(nums) at 1.04x faster than find(isZero, nums)

Running suite findIndex where [lib/bench/findIndex-where.bench.js]...
>> _.findIndex(objs, {x: []}) x 602,254 ops/sec ±3.02% (88 runs sampled)
>> findIndex(where({x: isEmpty}), objs) x 540,340 ops/sec ±1.90% (94 runs sampled)
>> findIndex(where({x: isEmpty}))(objs) x 518,215 ops/sec ±2.51% (88 runs sampled)
>> findIndexEmptyX(objs) x 2,334,276 ops/sec ±1.46% (95 runs sampled)
>> _.findIndex(objs, {y: false}) x 1,370,551 ops/sec ±1.57% (97 runs sampled)
>> findIndex(where({y: false}), objs) x 818,761 ops/sec ±3.20% (94 runs sampled)
>> findIndex(where({y: false}))(objs) x 769,975 ops/sec ±1.47% (94 runs sampled)
>> findIndexFalseY(objs) x 5,791,335 ops/sec ±3.66% (89 runs sampled)
Fastest test is findIndexFalseY(objs) at 2.5x faster than findIndexEmptyX(objs)

Running suite findIndex [lib/bench/findIndex.bench.js]...
>> _.findIndex(nums, isZero) x 5,268,609 ops/sec ±5.69% (79 runs sampled)
>> findIndex(isZero, nums) x 10,135,591 ops/sec ±5.12% (84 runs sampled)
>> findIndex(isZero)(nums) x 8,533,386 ops/sec ±3.62% (87 runs sampled)
>> findIndexZero(nums) x 12,397,905 ops/sec ±2.45% (91 runs sampled)
Fastest test is findIndexZero(nums) at 1.22x faster than findIndex(isZero, nums)

Running suite indexOf [lib/bench/indexof.bench.js]...
>> _.indexOf x 25,674,999 ops/sec ±0.48% (95 runs sampled)
>> indexOf(sq, nums) x 20,509,740 ops/sec ±1.90% (91 runs sampled)
>> indexOf(sq)(nums) x 15,760,174 ops/sec ±3.19% (95 runs sampled)
>> idxOf23(nums) x 24,409,933 ops/sec ±2.06% (91 runs sampled)
Fastest tests are _.indexOf,idxOf23(nums) at 1.05x faster than idxOf23(nums)

Running suite isSet [lib/bench/isset.bench.js]...
>> isSet(unsortedBag) x 23,518,856 ops/sec ±0.72% (96 runs sampled)
>> isSet(unsortedSet) x 5,962,096 ops/sec ±1.13% (100 runs sampled)
>> isSet(sortedBag) x 22,310,952 ops/sec ±2.09% (94 runs sampled)
>> isSet(sortedSet) x 6,006,676 ops/sec ±0.66% (97 runs sampled)
Fastest test is isSet(unsortedBag) at 1.05x faster than isSet(sortedBag)

Running suite map [lib/bench/map.bench.js]...
>> _.map x 4,448,616 ops/sec ±0.87% (99 runs sampled)
>> map(sq, nums) x 4,157,810 ops/sec ±1.32% (95 runs sampled)
>> map(sq)(nums) x 4,006,617 ops/sec ±0.83% (99 runs sampled)
>> mapSq(nums) x 4,345,454 ops/sec ±1.59% (96 runs sampled)
Fastest tests are _.map,mapSq(nums) at 1.02x faster than mapSq(nums)

Running suite maxBy [lib/bench/maxWith.bench.js]...
>> _.max x 149,081 ops/sec ±0.63% (101 runs sampled)
>> maxBy(computer, nums) x 412,335 ops/sec ±0.59% (100 runs sampled)
>> maxBy(computer)(vals) x 395,273 ops/sec ±1.30% (96 runs sampled)
>> maxVal(vals) x 405,991 ops/sec ±1.49% (98 runs sampled)
Fastest tests are maxBy(computer, nums),maxVal(vals) at 1.02x faster than maxVal(vals)

Running suite reduce [lib/bench/reduce.bench.js]...
>> _.reduce(nums, add, 0) x 4,279,931 ops/sec ±1.18% (98 runs sampled)
>> reduce(add, 0, nums) x 7,603,141 ops/sec ±1.67% (96 runs sampled)
>> reduce(add, 0)(nums) x 7,014,637 ops/sec ±0.95% (96 runs sampled)
>> reduceAdd(nums) x 9,120,461 ops/sec ±0.59% (99 runs sampled)
Fastest test is reduceAdd(nums) at 1.20x faster than reduce(add, 0, nums)

Done, without errors.
```